+++ inlang/source-code/cli/src/commands/lint/lint.test.ts
@@ -3,8 +3,8 @@ import { lintCommandAction } from "./index.js"
 import {
 	MessageLintRule,
 	Message,
-	ProjectConfig,
-	openInlangProject,
+	ProjectSettings,
+	loadProject,
 	Plugin,
 	type InlangModule,
 } from "@inlang/sdk"
@@ -55,7 +55,7 @@ const exampleMessages: Message[] = [
 	},
 ]
 
-async function setupInlang(enabledLintRule?: MessageLintRule) {
+async function setupProject(enabledLintRule?: MessageLintRule) {
 	const fs = createNodeishMemoryFs()
 
 	await fs.writeFile(
@@ -64,18 +64,14 @@ async function setupInlang(enabledLintRule?: MessageLintRule) {
 			sourceLanguageTag: "en",
 			languageTags: ["en", "de", "it"],
 			modules: ["_mockPlugin.js", ...(enabledLintRule ? ["lintRule.js"] : [])],
-			settings: {
-				"project.messageLintRuleLevels": {},
-			},
-		} satisfies ProjectConfig),
+			messageLintRuleLevels: {},
+		} satisfies ProjectSettings),
 	)
 
 	const _mockPlugin: Plugin = {
-		meta: {
-			id: "plugin.inlang.json",
-			description: { en: "Mock plugin description" },
-			displayName: { en: "Mock Plugin" },
-		},
+		id: "plugin.inlang.json",
+		description: { en: "Mock plugin description" },
+		displayName: { en: "Mock Plugin" },
 		loadMessages: () => exampleMessages,
 		saveMessages: () => undefined as any,
 	}
@@ -89,8 +85,8 @@ async function setupInlang(enabledLintRule?: MessageLintRule) {
 		}
 	}
 
-	return await openInlangProject({
-		projectFilePath: "./project.inlang.json",
+	return await loadProject({
+		settingsFilePath: "./project.inlang.json",
 		nodeishFs: fs,
 		_import,
 	})
@@ -99,17 +95,15 @@ async function setupInlang(enabledLintRule?: MessageLintRule) {
 describe("lint command", () => {
 	it("succeed on lint success", async () => {
 		const enabledLintRule: MessageLintRule = {
-			meta: {
-				id: "messageLintRule.namespace.enabled",
-				description: { en: "Mock lint rule description" },
-				displayName: { en: "Mock Lint Rule" },
-			},
-			message: () => {
+			id: "messageLintRule.namespace.enabled",
+			description: { en: "Mock lint rule description" },
+			displayName: { en: "Mock Lint Rule" },
+			run: () => {
 				;/ * no lint reports for this test case * /
 			},
 		}
 
-		const inlang = await setupInlang(enabledLintRule)
+		const project = await setupProject(enabledLintRule)
 
 		const logger = {
 			log: vi.fn(),
@@ -120,7 +114,7 @@ describe("lint command", () => {
 		let lintResult
 		try {
 			lintResult = await lintCommandAction({
-				inlang,
+				project,
 				logger,
 			})
 		} catch (err) {
@@ -136,12 +130,10 @@ describe("lint command", () => {
 
 	it("show lint reports", async () => {
 		const enabledLintRule: MessageLintRule = {
-			meta: {
-				id: "messageLintRule.namespace.enabled",
-				description: { en: "Mock lint rule description" },
-				displayName: { en: "Mock Lint Rule" },
-			},
-			message: ({ report }) => {
+			id: "messageLintRule.namespace.enabled",
+			description: { en: "Mock lint rule description" },
+			displayName: { en: "Mock Lint Rule" },
+			run: ({ report }) => {
 				report({
 					messageId: "some-message-1",
 					languageTag: "en",
@@ -150,7 +142,7 @@ describe("lint command", () => {
 			},
 		}
 
-		const inlang = await setupInlang(enabledLintRule)
+		const project = await setupProject(enabledLintRule)
 
 		const logger = {
 			log: vi.fn(),
@@ -161,7 +153,7 @@ describe("lint command", () => {
 		let lintResult
 		try {
 			lintResult = await lintCommandAction({
-				inlang,
+				project,
 				logger,
 			})
 		} catch (err) {
@@ -177,7 +169,7 @@ describe("lint command", () => {
 	})
 
 	it("error on missing lint rules", async () => {
-		const inlang = await setupInlang()
+		const project = await setupProject()
 
 		const logger = {
 			log: vi.fn(),
@@ -188,7 +180,7 @@ describe("lint command", () => {
 		let lintResult
 		try {
 			lintResult = await lintCommandAction({
-				inlang,
+				project,
 				logger,
 			})
 		} catch (err) {