+++ inlang/source-code/ide-extension/src/state.ts
@@ -7,7 +7,7 @@ type State = {
 	/**
 	 * Inlang project
 	 */
-	inlang: InlangProject
+	project: InlangProject
 }
 
 let _state: State