+++ inlang/source-code/sdk/src/lint/message/lintMessages.ts
@@ -1,19 +1,11 @@
 import type { Message } from "@inlang/message"
 import { lintSingleMessage } from "./lintSingleMessage.js"
 import type { MessagedLintRuleThrowedError } from "./errors.js"
-import type { LanguageTag } from "@inlang/language-tag"
-import type { JSONObject } from "@inlang/json-types"
-import type {
-	MessageLintLevel,
-	MessageLintReport,
-	MessageLintRule,
-} from "@inlang/message-lint-rule"
+import type { MessageLintReport, MessageLintRule } from "@inlang/message-lint-rule"
+import type { ProjectSettings } from "@inlang/project-settings"
 
 export const lintMessages = async (args: {
-	sourceLanguageTag: LanguageTag
-	languageTags: LanguageTag[]
-	ruleSettings: Record<MessageLintRule["meta"]["id"], JSONObject>
-	ruleLevels: Record<MessageLintRule["meta"]["id"], MessageLintLevel>
+	settings: ProjectSettings & Required<Pick<ProjectSettings, "messageLintRuleLevels">>
 	rules: MessageLintRule[]
 	messages: Message[]
 }): Promise<{ data: MessageLintReport[]; errors: MessagedLintRuleThrowedError[] }> => {