+++ inlang/source-code/sdk/src/lint/message/lintSingleMessage.test.ts
@@ -5,21 +5,17 @@ import type { Message } from "@inlang/message"
 import { tryCatch } from "@inlang/result"
 
 const lintRule1 = {
-	meta: {
-		id: "messageLintRule.r.1",
-		displayName: { en: "" },
-		description: { en: "" },
-	},
-	message: vi.fn(),
+	id: "messageLintRule.r.1",
+	displayName: "",
+	description: "",
+	run: vi.fn(),
 } satisfies MessageLintRule
 
 const lintRule2 = {
-	meta: {
-		id: "messageLintRule.r.2",
-		displayName: { en: "" },
-		description: { en: "" },
-	},
-	message: vi.fn(),
+	id: "messageLintRule.r.2",
+	displayName: "",
+	description: "",
+	run: vi.fn(),
 } satisfies MessageLintRule
 
 const message1 = {} as Message
@@ -35,14 +31,16 @@ describe("lintSingleMessage", async () => {
 		// the lint function is un-opinionated and does not set a default level.
 		// opinionated users like the inlang instance can very well set a default level (separation of concerns)
 		test("it should throw if a lint level is not provided for a given lint rule", async () => {
-			lintRule1.message.mockImplementation(({ report }) => report({} as MessageLintReport))
+			lintRule1.run.mockImplementation(({ report }) => report({} as MessageLintReport))
 
 			const result = await tryCatch(() =>
 				lintSingleMessage({
-					ruleLevels: {},
-					ruleSettings: {},
-					sourceLanguageTag: "en",
-					languageTags: ["en"],
+					settings: {
+						sourceLanguageTag: "en",
+						languageTags: ["en"],
+						messageLintRuleLevels: {},
+						modules: [],
+					},
 					messages,
 					message: message1,
 					rules: [lintRule1],
@@ -53,15 +51,17 @@ describe("lintSingleMessage", async () => {
 		})
 
 		test("it should override the default lint level", async () => {
-			lintRule1.message.mockImplementation(({ report }) => report({} as MessageLintReport))
+			lintRule1.run.mockImplementation(({ report }) => report({} as MessageLintReport))
 
 			const reports = await lintSingleMessage({
-				ruleLevels: {
-					[lintRule1.meta.id]: "error",
+				settings: {
+					sourceLanguageTag: "en",
+					languageTags: ["en"],
+					modules: [],
+					messageLintRuleLevels: {
+						[lintRule1.id]: "error",
+					},
 				},
-				ruleSettings: {},
-				sourceLanguageTag: "en",
-				languageTags: ["en"],
 				messages,
 				message: message1,
 				rules: [lintRule1],
@@ -70,20 +70,20 @@ describe("lintSingleMessage", async () => {
 		})
 
 		test("it should pass the correct settings", async () => {
-			const settings = {}
-
-			const fn = vi.fn()
-			lintRule1.message.mockImplementation(({ settings }) => fn(settings))
-
-			await lintSingleMessage({
-				ruleLevels: {
-					[lintRule1.meta.id]: "warning",
-				},
-				ruleSettings: {
-					[lintRule1.meta.id]: settings,
-				},
+			const settings = {
 				sourceLanguageTag: "en",
 				languageTags: ["en"],
+				modules: [],
+				messageLintRuleLevels: {
+					[lintRule1.id as any]: "warning",
+				},
+			}
+
+			const fn = vi.fn()
+			lintRule1.run.mockImplementation(({ settings }) => fn(settings))
+
+			await lintSingleMessage({
+				settings,
 				messages,
 				message: message1,
 				rules: [lintRule1],
@@ -96,22 +96,24 @@ describe("lintSingleMessage", async () => {
 	test("it should await all rules", async () => {
 		let m1Called = false
 		let m2Called = false
-		lintRule1.message.mockImplementation(() => {
+		lintRule1.run.mockImplementation(() => {
 			m1Called = true
 		})
-		lintRule2.message.mockImplementation(async () => {
+		lintRule2.run.mockImplementation(async () => {
 			await new Promise((resolve) => setTimeout(resolve, 0))
 			m2Called = true
 		})
 
 		await lintSingleMessage({
-			ruleLevels: {
-				[lintRule1.meta.id]: "warning",
-				[lintRule2.meta.id]: "warning",
+			settings: {
+				sourceLanguageTag: "en",
+				languageTags: ["en"],
+				modules: [],
+				messageLintRuleLevels: {
+					[lintRule1.id]: "warning",
+					[lintRule2.id]: "warning",
+				},
 			},
-			ruleSettings: {},
-			sourceLanguageTag: "en",
-			languageTags: ["en"],
 			messages,
 			message: message1,
 			rules: [lintRule1, lintRule2],
@@ -124,54 +126,58 @@ describe("lintSingleMessage", async () => {
 	test("it should process all rules in parallel", async () => {
 		const fn = vi.fn()
 
-		lintRule1.message.mockImplementation(async () => {
-			fn(lintRule1.meta.id, "before")
+		lintRule1.run.mockImplementation(async () => {
+			fn(lintRule1.id, "before")
 			await new Promise((resolve) => setTimeout(resolve, 0))
-			fn(lintRule1.meta.id, "after")
+			fn(lintRule1.id, "after")
 		})
-		lintRule2.message.mockImplementation(async () => {
-			fn(lintRule2.meta.id, "before")
+		lintRule2.run.mockImplementation(async () => {
+			fn(lintRule2.id, "before")
 			await new Promise((resolve) => setTimeout(resolve, 0))
-			fn(lintRule2.meta.id, "after")
+			fn(lintRule2.id, "after")
 		})
 
 		await lintSingleMessage({
-			ruleLevels: {
-				[lintRule1.meta.id]: "warning",
-				[lintRule2.meta.id]: "warning",
+			settings: {
+				sourceLanguageTag: "en",
+				languageTags: ["en"],
+				modules: [],
+				messageLintRuleLevels: {
+					[lintRule1.id]: "warning",
+					[lintRule2.id]: "warning",
+				},
 			},
-			ruleSettings: {},
-			sourceLanguageTag: "en",
-			languageTags: ["en"],
 			messages,
 			message: message1,
 			rules: [lintRule1, lintRule2],
 		})
 
 		expect(fn).toHaveBeenCalledTimes(4)
-		expect(fn).toHaveBeenNthCalledWith(1, lintRule1.meta.id, "before")
-		expect(fn).toHaveBeenNthCalledWith(2, lintRule2.meta.id, "before")
-		expect(fn).toHaveBeenNthCalledWith(3, lintRule1.meta.id, "after")
-		expect(fn).toHaveBeenNthCalledWith(4, lintRule2.meta.id, "after")
+		expect(fn).toHaveBeenNthCalledWith(1, lintRule1.id, "before")
+		expect(fn).toHaveBeenNthCalledWith(2, lintRule2.id, "before")
+		expect(fn).toHaveBeenNthCalledWith(3, lintRule1.id, "after")
+		expect(fn).toHaveBeenNthCalledWith(4, lintRule2.id, "after")
 	})
 
 	test("it should not abort the linting process when errors occur", async () => {
-		lintRule1.message.mockImplementation(() => {
+		lintRule1.run.mockImplementation(() => {
 			throw new Error("error")
 		})
 
-		lintRule2.message.mockImplementation(({ report }) => {
+		lintRule2.run.mockImplementation(({ report }) => {
 			report({} as MessageLintReport)
 		})
 
 		const result = await lintSingleMessage({
-			ruleLevels: {
-				[lintRule1.meta.id]: "warning",
-				[lintRule2.meta.id]: "warning",
+			settings: {
+				sourceLanguageTag: "en",
+				languageTags: ["en"],
+				modules: [],
+				messageLintRuleLevels: {
+					[lintRule1.id]: "warning",
+					[lintRule2.id]: "warning",
+				},
 			},
-			ruleSettings: {},
-			sourceLanguageTag: "en",
-			languageTags: ["en"],
 			messages,
 			message: message1,
 			rules: [lintRule1, lintRule2],