+++ inlang/source-code/templates/plugin/src/plugin.test.ts
@@ -1,0 +1,34 @@
+import { it, expect } from "vitest"
+import { ProjectSettings, loadProject } from "@inlang/sdk"
+import { createNodeishMemoryFs } from "@inlang/sdk/test-utilities"
+import { id as pluginId } from "../marketplace-manifest.json"
+
+it("should return fake messages to illustrate how a plugin works", async () => {
+	// creating a virtual filesystem to store the project file
+	const fs = createNodeishMemoryFs()
+
+	// creating a project file
+	const settings = {
+		sourceLanguageTag: "en",
+		modules: ["./plugin.js"],
+		languageTags: ["en", "de"],
+	} satisfies ProjectSettings
+
+	// writing the project file to the virtual filesystem
+	await fs.writeFile("/project.inlang.json", JSON.stringify(settings))
+
+	// opening the project file and loading the plugin
+	const project = await loadProject({
+		nodeishFs: fs,
+		settingsFilePath: "/project.inlang.json",
+		// simulate the import function that the SDK uses
+		// to inject the plugin into the project
+		_import: async () => import("./index.js"),
+	})
+
+	expect(project.errors()).toEqual([])
+
+	expect(project.installed.plugins()[0]?.id).toBe(pluginId)
+
+	expect(project.query.messages.get({ where: { id: "this-is-a-test-message" } })).toBeDefined()
+})