+++ inlang/source-code/cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @inlang/cli
 
+## 1.6.0
+
+### Minor Changes
+
+- 25fe8502: refactor: remove plugin.meta and messageLintRule.meta nesting
+
+## 1.5.0
+
+### Minor Changes
+
+- dcb20e90: improve: module init got new templates with testing and more
+
+## 1.4.0
+
+### Minor Changes
+
+- 81b9d741: refactor: move 'config validate' command to 'project validate'
+- d0592709: improve: simplify init template
+
+## 1.3.0
+
+### Minor Changes
+
+- 6ace5de0: update README of CLI to provide new commands `module` and `project`
+
 ## 1.2.0
 
 ### Minor Changes